% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidies.R, R/arrange.R, R/filter.R,
%   R/group-by.R, R/select.R, R/summarise.R, R/tibble.R
\docType{package}
\name{tidies-package}
\alias{tidies}
\alias{tidies-package}
\alias{arrange.eSet}
\alias{filter.eSet}
\alias{group_by.eSet}
\alias{select.eSet}
\alias{summarise.Grouped_eSet}
\alias{as_tibble.MSnSet}
\title{A Grammar of Data Manipulation for Quantitative Proteomics}
\usage{
\method{arrange}{eSet}(.data, ...)

\method{filter}{eSet}(.data, ...)

\method{group_by}{eSet}(.data, ..., add = FALSE)

\method{select}{eSet}(.data, ...)

\method{summarise}{Grouped_eSet}(.data, ...)

\method{as_tibble}{MSnSet}(x, ..., fcols = fvarLabels(x))
}
\arguments{
\item{.data}{An object of class \link[MSnbase:MSnSet]{MSnbase::MSnSet}.}

\item{...}{Expressions evaluated in the context of the object's
feature and sample variable and passed to the \code{dplyr}
functions. Ignored in \code{as_tibble,MSnSet}.}

\item{add}{As in the original `dplyr::group_by`function, when ‘add
= FALSE’, the default, `group_by()` will override existing
groups. To add to the existing groups, use `add = TRUE`.}

\item{x}{An object of class \code{MSnSet}.}

\item{fcols}{Feature variables to be added. Default is to add all
(i.e. \code{fvarLabels(x)}). Use \code{NULL} to add none.}
}
\description{
The `tidyms` package implements tidy principles as defined in the
tidyverse packages to omics-type data classes, with (currently at
least), an emphasis on quantitative proteomics data.
}
\details{
The vignette provides additional details and examples.

See the original `dplyr` manual pages for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lgatto/tidies}
  \item Report bugs at \url{https://github.com/lgatto/tidies/issues}
}

}
\author{
\strong{Maintainer}: Laurent Gatto \email{lg390@cam.ac.uk}

}
