% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyms.R, R/arrange.R, R/filter.R,
%   R/group-by.R, R/select.R, R/summarise.R
\docType{package}
\name{tidyms-package}
\alias{tidyms}
\alias{tidyms-package}
\alias{arrange.MSnSet}
\alias{filter.MSnSet}
\alias{group_by.MSnSet}
\alias{select.MSnSet}
\alias{summarise.GroupedMSnSet}
\title{A Grammar of Data Manipulation for Quantitative Proteomics}
\usage{
\method{arrange}{MSnSet}(.data, ...)

\method{filter}{MSnSet}(.data, ...)

\method{group_by}{MSnSet}(.data, ..., add = FALSE)

\method{select}{MSnSet}(.data, ...)

\method{summarise}{GroupedMSnSet}(.data, ...)
}
\arguments{
\item{.data}{An object of class [MSnbase::MSnSet].}

\item{...}{Expressions evaluated in the context of the object's
feature and sample variable and passed to the `dplyr`
functions.}

\item{add}{As in the original `dplyr::group_by`function, when ‘add
= FALSE’, the default, `group_by()` will override existing
groups. To add to the existing groups, use `add = TRUE`.}
}
\description{
The `tidyms` package implements tidy principles as defined in the
tidyverse packages to omics-type data classes, with (currently at
least), an emphasis on quantitative proteomics data.
}
\details{
The vignette provides additional details and examples.

See the original `dplyr` manual pages for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lgatto/tidyms}
  \item Report bugs at \url{https://github.com/lgatto/tidyms/issues}
}

}
\author{
\strong{Maintainer}: Laurent Gatto \email{lg390@cam.ac.uk}

}
