% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{class}
\name{Grouped_eSet-class}
\alias{Grouped_eSet-class}
\alias{.Grouped_eSet}
\alias{show,Grouped_eSet-method}
\title{A eSet/MSnSet object with grouped sample or feature variables}
\usage{
\S4method{show}{Grouped_eSet}(object)
}
\arguments{
\item{object}{An `Grouped_eSet` instance.}
}
\description{
This object extends the \link[MSnbase:MSnSet]{MSnbase::MSnSet} class by
adding slots necessary to record the grouping structure. These
additional slots are inherited from the \link[dplyr:grouped_df]{dplyr::grouped_df}
class.\preformatted{Objects of class `Grouped_eSet` are created by the `group_by`
function. See the vignette for examples.
}
}
\section{Slots}{

\describe{
\item{\code{fvars}}{a \code{character} vector or a \code{list} of grouping feature
variables.}

\item{\code{fdrop}}{a logical (default is \code{TRUE}) to preserve all feature
variable factor levels, even those without data.}

\item{\code{findices}}{indices defining the feature groups.}

\item{\code{fgroup_sizes}}{an \code{integer} defining the sizes of the feature
groups.}

\item{\code{fbiggest_group_size}}{an \code{integer} defining the biggest
feature group size.}

\item{\code{flabels}}{a \code{data.frame} defining the feature groups and
levels.}

\item{\code{pvars}}{a \code{character} vector or a \code{list} of grouping sample
variables.}

\item{\code{pdrop}}{a logical (default is \code{TRUE}) to preserve all sample
variable factor levels, even those without data.}

\item{\code{pindices}}{indices defining the sample groups.}

\item{\code{pgroup_sizes}}{an \code{integer} defining the sizes of the sample
groups.}

\item{\code{pbiggest_group_size}}{an \code{integer} defining the biggest sample
group size.}

\item{\code{plabels}}{a \code{data.frame} defining the sample groups and
levels.}
}}

